# web/public/api/openapi.yaml
openapi: 3.1.0
info:
  title: UNIS Fulfillment Quotation API
  version: 1.1.0
  summary: Customers, quotes, versions, rate sheets and charge-code mapback for UNIS warehouse services.
  description: |
    Integration API for the UNIS Fulfillment quotation tool.

    **Main objects** (full CRUD): `Customer`, `Quote`, the quote `Draft` and its per-charge `Selection`s,
    catalog `Charge`s, `ChargeCodeMapRule`s and `Webhook`s.
    **Versions** are immutable snapshots: create, read, restore. **History** is append-only.
    **Computed resources** are read-only: rate lines, the customer rate sheet (proposal), the charge-code mapback,
    setup gaps and version diffs.

    ### Charge-code mapback
    Every rate line of a quote resolves to an existing billing-system charge code plus the system conditions to configure
    (for example `HANDLING-0129` with `OFFLoad Type: Floor Loaded`, `Case Qty: 0-500`). Each line has a status:

    | Status | Meaning |
    | --- | --- |
    | `mapped` | The code exists and already uses every condition the line needs |
    | `new-condition` | The code exists, but a condition the line needs is not configured on it today |
    | `new-item` | No code exists yet; create the item in the billing system first (`suggestedName`) |

    `POST /quotes/{quoteId}/versions` stores the mapback with the version. Pass `requireMapped=true`
    to refuse a version that still has `new-condition` or `new-item` lines.

    ### Conventions
...

    ### Authentication
...

...
servers:
...
security:
  - cookieAuth: []
  - apiKey: []
...
paths:
...
  /quotes:
    get:
      tags: [Quotes]
      operationId: listQuotes
      summary: List quotes
      parameters:
        - { name: customerId, in: query, schema: { type: string } }
        - { name: status, in: query, schema: { $ref: "#/components/schemas/QuoteStatus" } }
        - { name: updatedSince, in: query, schema: { type: string, format: date-time } }
        - { name: q, in: query, description: Search quote number or title., schema: { type: string } }
        - $ref: "#/components/parameters/page"
        - $ref: "#/components/parameters/pageSize"
      responses:
        "401": { $ref: "#/components/responses/Unauthorized" }
        "200":
          description: Quotes (without draft selections; fetch the draft separately)
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Page"
                  - type: object
                    properties:
                      items: { type: array, items: { $ref: "#/components/schemas/QuoteSummary" } }
    post:
      tags: [Quotes]
      operationId: createQuote
      summary: Create a quote
      description: A new quote starts from the default template (the most common charges) unless `template` is `blank`.
      parameters: [{ $ref: "#/components/parameters/idempotencyKey" }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/QuoteCreate" }
      responses:
        "401": { $ref: "#/components/responses/Unauthorized" }
        "201":
          description: Created
          headers: { ETag: { $ref: "#/components/headers/ETag" }, Location: { schema: { type: string } } }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Quote" }
        "404": { description: Customer not found, content: { application/problem+json: { schema: { $ref: "#/components/schemas/Problem" } } } }
        "422": { $ref: "#/components/responses/Problem" }
  /quotes/{quoteId}:
    parameters: [{ $ref: "#/components/parameters/quoteId" }]
    get:
      tags: [Quotes]
      operationId: getQuote
      summary: Get a quote (header, status, draft and version list)
      responses:
        "401": { $ref: "#/components/responses/Unauthorized" }
        "200":
          description: Quote
          headers: { ETag: { $ref: "#/components/headers/ETag" } }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Quote" }
        "404": { $ref: "#/components/responses/Problem" }
    patch:
      tags: [Quotes]
      operationId: updateQuote
      summary: Update quote status or draft header
      parameters: [{ $ref: "#/components/parameters/ifMatch" }]
      requestBody:
        required: true
        content:
          application/merge-patch+json:
            schema: { $ref: "#/components/schemas/QuotePatch" }
      responses:
        "401": { $ref: "#/components/responses/Unauthorized" }
        "200": { description: Updated, content: { application/json: { schema: { $ref: "#/components/schemas/Quote" } } } }
        "404": { $ref: "#/components/responses/Problem" }
        "412": { $ref: "#/components/responses/Problem" }
        "422": { $ref: "#/components/responses/Problem" }
    delete:
      tags: [Quotes]
      operationId: deleteQuote
      summary: Delete a quote
      description: Quotes with saved versions are archived (status `archived`) instead of removed; drafts without versions are deleted.
      responses:
        "401": { $ref: "#/components/responses/Unauthorized" }
        "204": { description: Deleted or archived }
        "404": { $ref: "#/components/responses/Problem" }
  /quotes/{quoteId}/draft:
    parameters: [{ $ref: "#/components/parameters/quoteId" }]
    get:
      tags: [Drafts]
      operationId: getDraft
      summary: Get the draft (header + all charge selections)
      responses:
        "401": { $ref: "#/components/responses/Unauthorized" }
        "200":
          description: Draft
          headers: { ETag: { $ref: "#/components/headers/ETag" } }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/QuoteData" }
        "404": { $ref: "#/components/responses/Problem" }
    put:
      tags: [Drafts]
      operationId: replaceDraft
      summary: Replace the draft
      parameters:
        - $ref: "#/components/parameters/ifMatch"
        - { name: template, in: query, description: "Reset to a template instead of using the body: `default` or `blank`.", schema: { type: string, enum: [default, blank] } }
      requestBody:
        content:
          application/json:
            schema: { $ref: "#/components/schemas/QuoteData" }
      responses:
        "401": { $ref: "#/components/responses/Unauthorized" }
        "200": { description: Updated, content: { application/json: { schema: { $ref: "#/components/schemas/QuoteData" } } } }
        "404": { $ref: "#/components/responses/Problem" }
        "412": { $ref: "#/components/responses/Problem" }
        "422": { description: "Unknown charge, unit, factor value or driver", content: { application/problem+json: { schema: { $ref: "#/components/schemas/Problem" } } } }
  /quotes/{quoteId}/draft/selections:
    parameters: [{ $ref: "#/components/parameters/quoteId" }]
    get:
      tags: [Drafts]
      operationId: listSelections
      summary: List charge selections in the draft
      parameters:
        - { name: onlyOn, in: query, description: Only charges included in the quote., schema: { type: boolean, default: true } }
      responses:
        "401": { $ref: "#/components/responses/Unauthorized" }
        "200":
          description: Selections keyed by charge ID
          content:
            application/json:
              schema: { type: object, additionalProperties: { $ref: "#/components/schemas/ChargeSelection" } }
  /quotes/{quoteId}/draft/selections/{chargeId}:
    parameters: [{ $ref: "#/components/parameters/quoteId" }, { $ref: "#/components/parameters/chargeId" }]
    get:
      tags: [Drafts]
      operationId: getSelection
      summary: Get one charge selection
      responses:
        "401": { $ref: "#/components/responses/Unauthorized" }
        "200": { description: Selection, content: { application/json: { schema: { $ref: "#/components/schemas/ChargeSelection" } } } }
        "404": { $ref: "#/components/responses/Problem" }
    put:
      tags: [Drafts]
      operationId: putSelection
      summary: Create or replace one charge selection
      description: Validated against the catalog (unit, driver and factor values must exist; units hidden by the show/hide rules are rejected unless `showAll` is true).
      parameters: [{ $ref: "#/components/parameters/ifMatch" }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ChargeSelection" }
      responses:
        "401": { $ref: "#/components/responses/Unauthorized" }
        "200": { description: Saved, content: { application/json: { schema: { $ref: "#/components/schemas/ChargeSelection" } } } }
        "404": { $ref: "#/components/responses/Problem" }
        "412": { $ref: "#/components/responses/Problem" }
        "422": { $ref: "#/components/responses/Problem" }
    delete:
      tags: [Drafts]
      operationId: deleteSelection
      summary: Remove a charge from the draft
      responses:
        "401": { $ref: "#/components/responses/Unauthorized" }
        "204": { description: Removed }
        "404": { $ref: "#/components/responses/Problem" }
  /quotes/{quoteId}/versions:
    parameters: [{ $ref: "#/components/parameters/quoteId" }]
    get:
      tags: [Versions]
      operationId: listVersions
      summary: List saved versions
      responses:
        "401": { $ref: "#/components/responses/Unauthorized" }
        "200":
          description: Versions, oldest first (without snapshot data)
          content:
            application/json:
              schema: { type: array, items: { $ref: "#/components/schemas/VersionSummary" } }
    post:
      tags: [Versions]
      operationId: saveVersion
      summary: Save the current draft as a new version
      description: |
        Snapshots the draft and its charge-code mapback. With `requireMapped=true` the request fails with `409`
        (body lists the setup items) while any rate line is `new-condition` or `new-item`.
      parameters:
        - $ref: "#/components/parameters/idempotencyKey"
        - { name: requireMapped, in: query, schema: { type: boolean, default: false } }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                note: { type: string, maxLength: 2000, description: What changed. }
      responses:
        "401": { $ref: "#/components/responses/Unauthorized" }
        "201":
          description: Version saved
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Version" }
        "409":
          description: Setup required before the version can be saved (only with requireMapped=true)
          content:
            application/problem+json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Problem"
                  - type: object
                    properties:
                      setup: { type: array, items: { $ref: "#/components/schemas/SetupItem" } }
...
  /quotes/{quoteId}/compare:
    parameters: [{ $ref: "#/components/parameters/quoteId" }]
    get:
      tags: [Versions]
      operationId: compareVersions
      summary: Compare two versions, or a version with the draft
      parameters:
        - { name: from, in: query, required: true, description: "Version number or `draft`.", schema: { type: string } }
        - { name: to, in: query, required: true, description: "Version number or `draft`.", schema: { type: string } }
      responses:
        "401": { $ref: "#/components/responses/Unauthorized" }
        "200":
          description: Rate-line differences
          content:
            application/json:
              schema: { type: array, items: { $ref: "#/components/schemas/RateLineDiff" } }
  /quotes/{quoteId}/rate-lines:
    parameters: [{ $ref: "#/components/parameters/quoteId" }, { $ref: "#/components/parameters/versionQuery" }]
    get:
      tags: [Rate sheet]
      operationId: listRateLines
      summary: Every priced cell of the quote, with its charge-code mapping
      responses:
        "401": { $ref: "#/components/responses/Unauthorized" }
        "200":
          description: Rate lines in catalog order
          content:
            application/json:
              schema: { type: array, items: { $ref: "#/components/schemas/RateLine" } }
  /quotes/{quoteId}/proposal:
    parameters: [{ $ref: "#/components/parameters/quoteId" }, { $ref: "#/components/parameters/versionQuery" }]
    get:
      tags: [Rate sheet]
      operationId: getProposal
      summary: Customer rate sheet (Warehouse Services Proposal)
      description: |
        Laid out like the UNIS Business Proposal template: sections in template order, same-price variants merged into one row,
        factors shared by every row lifted into the service heading. Use `Accept` to choose the format.
      parameters:
        - $ref: "#/components/parameters/lang"
      responses:
        "401": { $ref: "#/components/responses/Unauthorized" }
        "200":
          description: Proposal
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Proposal" }
            application/pdf:
              schema: { type: string, format: binary }
            application/vnd.openxmlformats-officedocument.spreadsheetml.sheet:
              schema: { type: string, format: binary }
        "404": { $ref: "#/components/responses/Problem" }
...
components:
  securitySchemes:
    cookieAuth: { type: apiKey, in: cookie, name: uq_session, description: "Session of a person signed in with Microsoft Entra ID (/api/auth/login)." }
    apiKey: { type: apiKey, in: header, name: X-API-Key, description: Key for system-to-system integrations. }
  parameters:
...
    quoteId: { name: quoteId, in: path, required: true, schema: { type: string } }
    chargeId: { name: chargeId, in: path, required: true, schema: { type: string }, example: IN-OFFLOAD }
...
    versionQuery: { name: version, in: query, description: "Version number; omit for the current draft.", schema: { type: integer, minimum: 1 } }
    lang: { name: lang, in: query, schema: { $ref: "#/components/schemas/Lang" } }
    page: { name: page, in: query, schema: { type: integer, minimum: 1, default: 1 } }
    pageSize: { name: pageSize, in: query, schema: { type: integer, minimum: 1, maximum: 200, default: 50 } }
    ifMatch: { name: If-Match, in: header, description: ETag from the last read; required for updates., schema: { type: string } }
    idempotencyKey: { name: Idempotency-Key, in: header, description: Unique key per logical request (UUID)., schema: { type: string, maxLength: 128 } }
  headers:
    ETag: { description: Version tag of the resource for If-Match., schema: { type: string } }
  responses:
    Unauthorized:
      description: Missing or invalid credentials
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    Problem:
      description: Error
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
  schemas:
    Problem:
      type: object
      description: RFC 9457 problem details.
      required: [title, status]
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
        errors:
          type: array
          items: { type: object, properties: { path: { type: string }, message: { type: string } } }
    Page:
      type: object
      required: [page, pageSize, total]
      properties:
        page: { type: integer }
        pageSize: { type: integer }
        total: { type: integer }
...
    QuoteStatus: { type: string, enum: [draft, sent, accepted, archived] }
...
    QuoteHeader:
      type: object
      properties:
        title: { type: string }
        facility: { type: string, description: "UNIS facility, empty = all facilities.", examples: ["Buena Park, CA"] }
        effectiveDate: { type: string, format: date }
        validDays: { type: integer, minimum: 1, default: 90 }
        preparedBy: { type: string }
        notes: { type: string }
...
    ChargeSelection:
      type: object
      required: [on]
      properties:
        on: { type: boolean, description: Charge included in the quote. }
        price: { type: [number, "null"], description: Simple charges only; omitted = benchmark default. }
        conds: { type: object, additionalProperties: { $ref: "#/components/schemas/CondSelection" } }
        units:
          type: object
          description: "Keyed by unit ID; a unit priced a second way uses `unitId#2`, `unitId#3`, ... (e.g. `case` flat and `case#2` by case weight)."
          additionalProperties: { $ref: "#/components/schemas/UnitSelection" }
        prices:
          type: object
          description: "Builder rate cells: key `unitId|value|value...|col` (col = p, first, add) or `unitId|MIN|basis`; omitted cells use the benchmark."
          additionalProperties: { type: [number, "null"] }
        settings: { type: object, additionalProperties: { type: string }, description: "Billing settings, e.g. initial storage." }
        showAll: { type: boolean }
      example:
        on: true
        conds: { offloadType: { on: true, values: [Floor loaded, Palletized] } }
        units: { container: { on: true, driver: caseCount }, pallet: { on: true, min: load } }
        prices: { "container|Floor loaded|0 - 500 cases|p": 495 }
    QuoteData:
      type: object
      required: [header, selections]
      properties:
        header: { $ref: "#/components/schemas/QuoteHeader" }
        selections: { type: object, additionalProperties: { $ref: "#/components/schemas/ChargeSelection" } }
    QuoteCreate:
      type: object
      required: [customerId]
      properties:
        customerId: { type: string }
        template: { type: string, enum: [default, blank], default: default }
        header: { $ref: "#/components/schemas/QuoteHeader" }
    QuotePatch:
      type: object
      properties:
        status: { $ref: "#/components/schemas/QuoteStatus" }
        header: { $ref: "#/components/schemas/QuoteHeader" }
    QuoteSummary:
      type: object
      required: [id, number, customerId, status, createdAt, updatedAt]
      properties:
        id: { type: string }
        number: { type: string, examples: [Q-2026-0001] }
        customerId: { type: string }
        title: { type: string }
        status: { $ref: "#/components/schemas/QuoteStatus" }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        latestVersion: { type: [integer, "null"] }
        draftChanged: { type: boolean, description: Draft differs from the latest version. }
    Quote:
      allOf:
        - $ref: "#/components/schemas/QuoteSummary"
        - type: object
          required: [draft, versions]
          properties:
            draft: { $ref: "#/components/schemas/QuoteData" }
            versions: { type: array, items: { $ref: "#/components/schemas/VersionSummary" } }
...
    VersionSummary:
      type: object
      required: [v, savedAt, lineCount]
      properties:
        v: { type: integer, minimum: 1 }
        savedAt: { type: string, format: date-time }
        savedBy: { type: string }
        note: { type: string }
        lineCount: { type: integer }
        mapping: { $ref: "#/components/schemas/MappingSummary" }
    Version:
      allOf:
        - $ref: "#/components/schemas/VersionSummary"
        - type: object
          required: [data, mappingLines]
          properties:
            data: { $ref: "#/components/schemas/QuoteData" }
            mappingLines: { type: array, items: { $ref: "#/components/schemas/LineMapping" } }
            setup: { type: array, items: { $ref: "#/components/schemas/SetupItem" } }
...
    RateLineDiff:
      type: object
      required: [kind, key]
      properties:
        kind: { type: string, enum: [added, removed, changed] }
        key: { type: string }
        label: { type: string }
        from: { type: [number, "null"] }
        to: { type: [number, "null"] }
...
    SetupItem:
      type: object
      required: [kind, chargeId, lineCount]
      properties:
        kind: { type: string, enum: [new-item, new-condition] }
        chargeId: { type: string }
        code: { type: [string, "null"] }
        systemName: { type: string }
        suggestedName: { type: string }
        conditionKey: { type: string }
        values: { type: array, items: { type: string } }
        lineCount: { type: integer }
...